Move storage benchmark to bench
