+++ .gitignore
@@ -64,9 +64,9 @@ contracts/artifacts
 integration_test/dapp_tests/artifacts
 
 # DB benchmark data and artifacts
-sei-db/state_db/bench/cryptosim/data/**
-sei-db/state_db/bench/cryptosim/bin/
-sei-db/state_db/bench/cryptosim/logs/
-sei-db/ledger_db/block/blocksim/bin/
+sei-db/bench/cryptosim/data/**
+sei-db/bench/cryptosim/bin/
+sei-db/bench/cryptosim/logs/
+sei-db/bench/blocksim/bin/
 sei-db/seiwal/walsim/bin/
 sei-db/db_engine/litt/bin/
