+++ sei-db/bench/blocksim/Makefile
@@ -1,12 +1,12 @@
-BINARY := bin/cryptosim
+BINARY := bin/blocksim
 LOGGER_BINARY := bin/configure-logger
 MODULE_ROOT ?= $(shell git -C "$(CURDIR)" rev-parse --show-toplevel)
-CMD_DIR := ./sei-db/state_db/bench/cryptosim/cmd
+CMD_DIR := ./sei-db/bench/blocksim/cmd
 
 .PHONY: build
 build:
 	@mkdir -p "$(CURDIR)/bin"
-	cd "$(MODULE_ROOT)" && go build -o "$(CURDIR)/$(BINARY)" "$(CMD_DIR)/cryptosim"
+	cd "$(MODULE_ROOT)" && go build -o "$(CURDIR)/$(BINARY)" "$(CMD_DIR)/blocksim"
 	cd "$(MODULE_ROOT)" && go build -o "$(CURDIR)/$(LOGGER_BINARY)" "$(CMD_DIR)/configure-logger"
 
 .PHONY: clean
