+++ sei-db/bench/blocksim/cmd/configure-logger/main.go
@@ -13,8 +13,8 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/sei-protocol/sei-chain/sei-db/bench/blocksim"
 	"github.com/sei-protocol/sei-chain/sei-db/common/utils"
-	"github.com/sei-protocol/sei-chain/sei-db/ledger_db/block/blocksim"
 )
 
 func main() {
