+++ sei-db/bench/cryptosim/cmd/cryptosim/main.go
@@ -13,7 +13,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/collectors"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"github.com/sei-protocol/sei-chain/sei-db/state_db/bench/cryptosim"
+	"github.com/sei-protocol/sei-chain/sei-db/bench/cryptosim"
 	"go.opentelemetry.io/otel"
 	otelprometheus "go.opentelemetry.io/otel/exporters/prometheus"
 	otelmetric "go.opentelemetry.io/otel/sdk/metric"
