+++ sei-db/bench/cryptosim/config/reciept-store.json
