+++ sei-db/bench/cryptosim/config/ss-composite-config.json
