+++ sei-db/bench/cryptosim/config/ss-composite-pebbledb-write-only.json
