+++ sei-db/bench/cryptosim/config/ss-composite-rocksdb-write-only.json
