+++ sei-db/bench/cryptosim/cryptosim.sh
