+++ sei-db/bench/cryptosim/util.go
